VM Ownership code unification for Infra/cloud
